Extract MessagesPage data loading to dataAccess module
