+++ webapp/src/dataAccess.ts
@@ -2,6 +2,7 @@ import { Cache } from '@/Cache';
 import MessageAdapterFactory from '@/utils/adapters/MessageAdapterFactory';
 import { RepoGit } from '@/RepoGit';
 import { ServerConfig, ServerProjectConfig } from '@/utils/serverConfig';
+import { getTranslationsIdText } from './utils/translationObjectUtil';
 
 export async function accessProjects() {
   const serverConfig = await ServerConfig.read();
@@ -21,6 +22,37 @@ export async function accessProject(name: string) {
   }
 
   return readProject(project);
+}
+
+export async function accessLanguage(
+  projectName: string,
+  languageName: string,
+) {
+  const serverConfig = await ServerConfig.read();
+  const project = serverConfig.projects.find(
+    (project) => project.name === projectName,
+  );
+
+  if (!project) {
+    return null;
+  }
+
+  await RepoGit.cloneIfNotExist(project);
+  const repoGit = await RepoGit.getRepoGit(project);
+  const lyraConfig = await repoGit.getLyraConfig();
+  const projectConfig = lyraConfig.getProjectConfigByPath(project.projectPath);
+  const msgAdapter = MessageAdapterFactory.createAdapter(projectConfig);
+  const messages = await msgAdapter.getMessages();
+  const translationsWithFilePath = await Cache.getLanguage(
+    projectName,
+    languageName,
+  );
+  const translations = getTranslationsIdText(translationsWithFilePath);
+
+  return {
+    messages,
+    translations,
+  };
 }
 
 async function readProject(project: ServerProjectConfig) {
